Removed classes from old system
