+++ core/src/thepaperpilot/strange/Main.java
@@ -7,7 +7,10 @@ import com.badlogic.gdx.assets.AssetManager;
 import com.badlogic.gdx.audio.Music;
 import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.graphics.GL20;
+import com.badlogic.gdx.graphics.g2d.ParticleEffect;
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.g2d.TextureAtlas;
+import com.badlogic.gdx.math.Matrix4;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
@@ -19,16 +22,15 @@ import java.util.ArrayList;
 
 public class Main extends Game implements Screen {
     public static final AssetManager manager = new AssetManager();
-    public static final ArrayList<Item> inventory = new ArrayList<Item>();
-    public static final ArrayList<Item> selected = new ArrayList<Item>();
     public static Skin skin;
     public static TextureAtlas animations;
     public static TextureAtlas backgrounds;
     public static TextureAtlas entities;
-    public static int[] decisions = new int[3];
     public static boolean reverse;
     private static Main instance;
+    private static SpriteBatch batch;
     private Stage loadingStage;
+    private static ParticleEffect choicesParticle;
 
     public static void changeScreen(Screen screen) {
         if (screen == null)
@@ -39,6 +41,15 @@ public class Main extends Game implements Screen {
     @Override
     public void create() {
         instance = this;
+
+        choicesParticle = new ParticleEffect();
+        choicesParticle.load(Gdx.files.internal("swirls.p"), Gdx.files.internal(""));
+        choicesParticle.setPosition(320, 180);
+        for (int i = 0; i < 10; i++) {
+            choicesParticle.update(.1f);
+        }
+
+        batch = new SpriteBatch();
 
         manager.load("skin.json", Skin.class);
 
@@ -80,14 +91,20 @@ public class Main extends Game implements Screen {
             entities = manager.get("entities.atlas", TextureAtlas.class);
             skin.getFont("large").getData().setScale(.5f);
             skin.getFont("font").getData().setScale(.25f);
-            Item.combine(); //trick to instantiate all the items
-            inventory.add(Item.NOTEBOOK);
-            Scene.updateInventory();
             manager.get("audio/bgm.ogg", Music.class).setLooping(true);
             manager.get("audio/bgm.ogg", Music.class).play();
 
             setScreen(new MenuScreen());
         }
+    }
+
+    public static void renderParticles(float delta) {
+        final Matrix4 trans = new Matrix4();
+        trans.scale(Gdx.graphics.getWidth() / 640, Gdx.graphics.getHeight() / 360, 1);
+        batch.setTransformMatrix(trans);
+        batch.begin();
+        choicesParticle.draw(batch, delta);
+        batch.end();
     }
 
     @Override
@@ -122,6 +139,7 @@ public class Main extends Game implements Screen {
         }
         manager.dispose();
         skin.dispose();
+        batch.dispose();
     }
 
     @Override
